requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
dj-database-url==1.2.0
Django==3.2.7
django-cors-headers==3.9.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
idna==3.4
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
whitenoise==6.3.0